extract utils; causes circular dependency
